Handle array jobs
The last index in an array job have the same jobid as the full array
so sacct prints out multiple lines. Include JobIDRaw in output and
pick the correct line.

This will also give a better error if the jobid is not found by sacct
at all, as sacct will return 0 with no output rather than fail.